ExpectedDocuments builder using radar-commons classes
